feat: Refactor network configuration to use domain_api and enhance gas fee module
