+++ crates/kari-move/Cargo.toml
@@ -29,6 +29,7 @@ serde_json.workspace = true
 hex.workspace = true
 sha3.workspace = true
 dirs.workspace = true
+common.workspace = true
 
 mona-types.workspace = true
 mona-vm.workspace = true
