+++ monaos/mona-types/src/lib.rs
@@ -2,3 +2,5 @@
 pub mod address;
 
 pub mod kari;
+
+pub mod gas;
